Moved actors to separate modules.
