+++ packages/jupyter-ai/jupyter_ai/actors/__init__.py
@@ -1,0 +1,1 @@
+"""Actor classes that process incoming chat messages"""
