+++ packages/jupyter-ai/pyproject.toml
@@ -29,6 +29,7 @@ dependencies = [
     "aiosqlite~=0.18",
     "importlib_metadata~=5.2.0",
     "langchain~=0.0.115",
+    "tiktoken", # required for OpenAIEmbeddings
     "jupyter_ai_magics",
     "ray==2.2.0", # latest ray version 2.3.0 requires grpcio installation from conda
     "faiss-cpu", # Not distributed by official repo
